// src/interfaces/IDice.ts
export type IDice = [
  {
    color: 'white';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  },
  {
    color: 'white';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  },
  {
    color: 'red';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  },
  {
    color: 'yellow';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  },
  {
    color: 'green';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  },
  {
    color: 'blue';
    value: 1 | 2 | 3 | 4 | 5 | 6;
  }
];